messaging-service/requirements.txt
# FastAPI et serveur web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# WebSockets
websockets==12.0

# MongoDB et ODM
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# Validation et configuration
pydantic==2.5.0
pydantic-settings==2.0.3

# HTTP client pour intégration services
httpx==0.25.2
requests==2.31.0

# Authentification et sécurité
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitaires
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Email et notifications
emails==0.6.0
jinja2==3.1.2

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Développement
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Logging et monitoring
structlog==23.2.0
